backend/requirements.txt
flask
transformers
torch
sentencepiece
PyPDF2
python-docx
